requirements.txt
# requirements.txt
Django==4.2.7
mysqlclient==2.2.0
Pillow==10.0.1
django-crispy-forms==2.0
crispy-bootstrap4==2022.1
python-dotenv==1.0.0
mysql-connector-python==8.2.0
google-generativeai
openai==1.3.0
razorpay==1.3.0